+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.14.11
+
+* Updates internal Pigeon API to use newer features.
+
+## 2.14.10
+
+* Adds 'PlatformBitmap' type.
+* Updates type unsafe implementations.
+
 ## 2.14.9
 
 * Adds `PlatformCap` for `PlatformPolyline.startCap` and `endCap`.